feat: add delete confirmation swal with sweetalert2

+++ index.js
@@ -6,6 +6,8 @@ const searchInput = document.getElementById("search-input");
 const deleteConfirmModal = document.getElementById("delete-contact-modal");
 const deleteConfirmButton = document.getElementById("confirm-delete-button");
 
+// const Swal = require("sweetalert2");
+
 // Save contacts to local storage
 const saveContacts = (contacts) => {
   try {
@@ -52,7 +54,7 @@ const renderContacts = (specificRenderContacts) => {
                     </svg>
                     Edit
                 </button>
-                <button id="delete-contact-button" type="click" contact-id="${contact.id}" data-modal-target="delete-contact-modal" data-modal-toggle="delete-contact-modal"
+                <button id="delete-contact-button" type="click" contact-id="${contact.id}"
                     class="flex items-center text-red-700 hover:text-white border border-red-700 hover:bg-red-800 focus:ring-4 focus:outline-none focus:ring-red-300 font-medium rounded-lg text-sm px-3 py-2 text-center dark:border-red-500 dark:text-red-500 dark:hover:text-white dark:hover:bg-red-600 dark:focus:ring-red-900">
                     <svg xmlns="http://www.w3.org/2000/svg" class="h-4 w-4 mr-1 -ml-0.5" viewbox="0 0 20 20" fill="currentColor"
                         aria-hidden="true">
@@ -110,8 +112,29 @@ contactList.addEventListener("click", (event) => {
     const contactId = parseInt(
       buttonDeleteTarget.getAttribute("contact-id", 10)
     );
-    const deletedContact = deleteContact(contactId);
-    renderContacts(deletedContact);
+
+    Swal.fire({
+      title: "Are you sure you want to delete this item?",
+      cancelButtonText: `Cancel`,
+      confirmButtonText: "Yes, I'm sure!",
+      showCancelButton: true,
+      confirmButtonColor: "#808080",
+      cancelButtonColor: "#d33",
+    }).then((result) => {
+      if (result.isConfirmed) {
+        Swal.fire({
+          title: "Deleted!",
+          text: "The contact has been deleted.",
+          icon: "success",
+          confirmButtonText: "OK!",
+          confirmButtonColor: "#0080fe",
+        });
+        const deletedContact = deleteContact(contactId);
+        renderContacts(deletedContact);
+      } else if (result.isDenied) {
+        Swal.fire("Cancel from deletion", "", "info");
+      }
+    });
   }
 });
